# public/swagger.yaml
components:
    examples: {}
    headers: {}
    parameters: {}
    requestBodies: {}
    responses: {}
    schemas:
        SignupResponse:
            properties:
                code:
                    type: number
                    format: double
                message:
                    type: string
            required:
                - code
                - message
            type: object
            additionalProperties: false
        LoginResponse:
            properties:
                code:
                    type: number
                    format: double
                message:
                    type: string
                refreshToken:
                    type: string
                accessToken:
                    type: string
            required:
                - code
                - message
                - refreshToken
                - accessToken
            type: object
            additionalProperties: false
        SessionResponse:
            properties:
                _id:
                    type: string
                name:
                    type: string
                email:
                    type: string
            required:
                - _id
                - name
                - email
            type: object
            additionalProperties: false
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
info:
    title: express-typescript
    version: 1.0.0
    license:
        name: ISC
    contact: {}
openapi: 3.0.0
paths:
    /api/auth/signup:
        post:
            operationId: Signup
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SignupResponse'
                            examples:
                                'Example 1':
                                    value: {code: 200, message: 'Signed Up Successfully'}
            tags:
                - Auth
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                email:
                                    type: string
                                password:
                                    type: string
                            required:
                                - name
                                - email
                                - password
    /api/auth/login:
        post:
            operationId: Login
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LoginResponse'
                            examples:
                                'Example 1':
                                    value: {accessToken: 'JWT Token', refreshToken: 'JWT Token', code: 200, message: 'Signed Up Successfully'}
            tags:
                - Auth
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                email:
                                    type: string
                                password:
                                    type: string
                            required:
                                - email
                                - password
    /api/auth/email/verify:
        post:
            operationId: VerifyEmail
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    message: {type: string}
                                    code: {type: number, format: double}
                                required:
                                    - message
                                    - code
                                type: object
                            examples:
                                'Example 1':
                                    value: {code: 200, message: 'Your account is verified successfully'}
            tags:
                - Auth
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                email:
                                    type: string
                                verificationCode:
                                    type: string
                            required:
                                - email
                                - verificationCode
    /api/session:
        get:
            operationId: GetSessionInfo
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SessionResponse'
                            examples:
                                'Example 1':
                                    value: {_id: '3928439', name: 'HS Rajput', email: email@gmail.com}
            tags:
                - Session
            security:
                -
                    bearerAuth: []
            parameters: []
servers:
    -
        url: /
tags:
    -
        name: Auth
        description: 'Operations about auth'
    -
        name: Session
        description: 'Operations about session'